+++ QiniuSDK/Utils/QNUtils.m
@@ -111,11 +111,11 @@
             type = [NSString stringWithFormat:@"%@-%ld", @"ipv4-A", (long)firstNumber];
         } else if (firstNumber > 127 && firstNumber <= 191) {
             NSInteger secondNumber = [ipNumberStrings[1] integerValue];
-            type = [NSString stringWithFormat:@"%@-%ld-%ld", @"ipv4-B", (long)firstNumber, secondNumber];
+            type = [NSString stringWithFormat:@"%@-%ld-%ld", @"ipv4-B", (long)firstNumber, (long)secondNumber];
         } else if (firstNumber > 191 && firstNumber <= 223) {
             NSInteger secondNumber = [ipNumberStrings[1] integerValue];
             NSInteger thirdNumber = [ipNumberStrings[2] integerValue];
-            type = [NSString stringWithFormat:@"%@-%ld-%ld-%ld", @"ipv4-C", (long)firstNumber, secondNumber, thirdNumber];
+            type = [NSString stringWithFormat:@"%@-%ld-%ld-%ld", @"ipv4-C", (long)firstNumber, (long)secondNumber, (long)thirdNumber];
         }
     }
     type = [NSString stringWithFormat:@"%@-%@", host ?:@"", type];
